refactor: apply clippy --fix and manual fixes to reduce warnings 156→77
- Ran cargo clippy --fix on all library targets
- Fixed 5 match arms with identical bodies (collection, lang, meta, relation)
- Converted needless_pass_by_value params to references in lang.rs
- Auto-fixed let...else, format strings, clone patterns, and doc issues

Progress: 156 clippy errors → 77 remaining (50% reduction)

Remaining issues are mostly architectural:
- 25 "large Err-variant" — Result types with oversized error types
- 10 "format appending" — String concatenation patterns
- 8 clone replacements — value type ownership
- 3 missing doc sections (docs for Result)
- Others: casting, doc formatting, pattern matching

These require structural refactoring beyond auto-fix capability.

+++ legend-pure-rust/crates/runtime/src/heap.rs
@@ -179,8 +179,8 @@ pub enum HeapEntry {
 /// via `Arc`.
 pub struct RuntimeHeap {
     objects: SlotMap<ObjectId, HeapEntry>,
-    /// BiMap between compiled `ElementId`s and the heap rows that
-    /// represent them as M3 metamodel CoreInstances. Populated by
+    /// `BiMap` between compiled `ElementId`s and the heap rows that
+    /// represent them as M3 metamodel `CoreInstances`. Populated by
     /// [`Self::bootstrap_metamodel`] at Evaluator construction;
     /// drives the unified-reflection refactor where every metamodel
     /// reference is reachable as both an `ElementId` (compiled-model
@@ -192,7 +192,7 @@ pub struct RuntimeHeap {
     /// reflective property access. The reverse direction
     /// (`object_to_element`) is rare; we synthesise it by walking
     /// the forward map at access time rather than storing two
-    /// HashMaps.
+    /// `HashMaps`.
     element_to_object: HashMap<ElementId, ObjectId>,
 }
 
@@ -518,7 +518,7 @@ impl RuntimeHeap {
         // `global_packages`, not in element chunks.
         for raw_pkg_idx in 0..model.global_packages.len() {
             #[allow(clippy::cast_possible_truncation)]
-            let pkg_id = PackageId(raw_pkg_idx as u32);
+            let pkg_id = PackageId(raw_pkg_idx);
             let eid = ElementId::Package(pkg_id);
             if self.element_to_object.contains_key(&eid) {
                 continue;
